add option to use pure Go ZSTD implementation for building application
